+++ clang-tools-extra/clang-tidy/misc/MiscTidyModule.cpp
@@ -32,6 +32,7 @@
 #include "UnusedUsingDeclsCheck.h"
 #include "UseAnonymousNamespaceCheck.h"
 #include "UseInternalLinkageCheck.h"
+#include "VisibilityChangeToVirtualFunctionCheck.h"
 
 namespace clang::tidy {
 namespace misc {
@@ -81,6 +82,8 @@ public:
         "misc-use-anonymous-namespace");
     CheckFactories.registerCheck<UseInternalLinkageCheck>(
         "misc-use-internal-linkage");
+    CheckFactories.registerCheck<VisibilityChangeToVirtualFunctionCheck>(
+        "misc-visibility-change-to-virtual-function");
   }
 };