+++ clang-tools-extra/clang-tidy/misc/VisibilityChangeToVirtualFunctionCheck.h
@@ -1,4 +1,4 @@
-//===--- FunctionVisibilityChangeCheck.h - clang-tidy -----------*- C++ -*-===//
+//===--- VisibilityChangeToVirtualFunctionCheck.h - clang-tidy --*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,23 +6,24 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BUGPRONE_FUNCTIONVISIBILITYCHANGECHECK_H
-#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BUGPRONE_FUNCTIONVISIBILITYCHANGECHECK_H
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISC_VISIBILITYCHANGETOVIRTUALFUNCTIONCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISC_VISIBILITYCHANGETOVIRTUALFUNCTIONCHECK_H
 
 #include "../ClangTidyCheck.h"
 
-namespace clang::tidy::bugprone {
+namespace clang::tidy::misc {
 
 /// Finds virtual function overrides with different visibility than the function
 /// in the base class.
 ///
 /// For the user-facing documentation see:
-/// http://clang.llvm.org/extra/clang-tidy/checks/bugprone/visibility-change-to-virtual-function.html
-class FunctionVisibilityChangeCheck : public ClangTidyCheck {
+/// http://clang.llvm.org/extra/clang-tidy/checks/misc/visibility-change-to-virtual-function.html
+class VisibilityChangeToVirtualFunctionCheck : public ClangTidyCheck {
 public:
   enum class ChangeKind { Any, Widening, Narrowing };
 
-  FunctionVisibilityChangeCheck(StringRef Name, ClangTidyContext *Context);
+  VisibilityChangeToVirtualFunctionCheck(StringRef Name,
+                                         ClangTidyContext *Context);
   void storeOptions(ClangTidyOptions::OptionMap &Opts) override;
   void registerMatchers(ast_matchers::MatchFinder *Finder) override;
   void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
@@ -37,6 +38,6 @@ private:
   std::vector<llvm::StringRef> IgnoredFunctions;
 };
 
-} // namespace clang::tidy::bugprone
+} // namespace clang::tidy::misc
 
-#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_BUGPRONE_FUNCTIONVISIBILITYCHANGECHECK_H
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISC_VISIBILITYCHANGETOVIRTUALFUNCTIONCHECK_H